Online part is working entirely
- Fixed some crashes, cuz some subclasses of model weren't Serializable too
- UI fixed
- App Icon
- BinCode type changed from Int to Long
- etc.

+++ app/src/main/java/com/cj3dreams/binchecker/repo/DataRepository.kt
@@ -4,6 +4,6 @@ import com.cj3dreams.binchecker.model.response.BinResponseModel
 import com.cj3dreams.binchecker.source.remote.RequestResult
 
 interface DataRepository {
-    suspend fun checkBin(bin: Int): RequestResult<BinResponseModel>
+    suspend fun checkBin(bin: Long): RequestResult<BinResponseModel>
     suspend fun getAllHistoryFromLocal()
 }
