+++ app/src/main/java/com/cj3dreams/binchecker/repo/DataRepositoryImpl.kt
@@ -5,7 +5,7 @@ import com.cj3dreams.binchecker.source.remote.RemoteSource
 import com.cj3dreams.binchecker.source.remote.RequestResult
 
 class DataRepositoryImpl(private val remoteSource: RemoteSource): DataRepository {
-    override suspend fun checkBin(bin: Int): RequestResult<BinResponseModel> =
+    override suspend fun checkBin(bin: Long): RequestResult<BinResponseModel> =
         remoteSource.checkBin(bin)
 
     override suspend fun getAllHistoryFromLocal() {
