+++ app/src/main/java/com/cj3dreams/binchecker/source/remote/RemoteSource.kt
@@ -3,5 +3,5 @@ package com.cj3dreams.binchecker.source.remote
 import com.cj3dreams.binchecker.model.response.BinResponseModel
 
 interface RemoteSource {
-    suspend fun checkBin(bin: Int): RequestResult<BinResponseModel>
+    suspend fun checkBin(bin: Long): RequestResult<BinResponseModel>
 }
