+++ app/src/main/java/com/cj3dreams/binchecker/usecase/CheckBinUseCase.kt
@@ -3,5 +3,5 @@ package com.cj3dreams.binchecker.usecase
 import com.cj3dreams.binchecker.repo.DataRepository
 
 class CheckBinUseCase (private val dataRepository: DataRepository) {
-    suspend operator fun invoke(bin: Int) = dataRepository.checkBin(bin)
+    suspend operator fun invoke(bin: Long) = dataRepository.checkBin(bin)
 }
